# docs/04-API/openapi.yaml
openapi: 3.0.3
info:
  title: 教务管理系统 API
  description: |
    培训机构**纯员工端 CRM** 后端 API（Express 4 + node:sqlite）。
    仅 `admin` / `teacher` 两种员工角色；学生、家长无账号、不登录。

    本文件是机器可读规范，人类可读的完整说明见 `docs/04-API/API.md`。
    覆盖范围：认证 + 四个 P0 模块（考勤签到 / 学生 / 成绩 / 招生线索）+ 通知 + AI 分析端点 + 课次实体与任课关系（v21）。
    统一响应包裹：成功 `{ success: true, data }`；失败 `{ success: false, message }`。
    `/api/analytics/*` 使用独立的「AI 分析信封」。
    v21（2026-09-23）：新增 `/api/sessions/*`（课次）、`/api/teaching-assignments`（任课关系）、`/api/period-times`（节次时间）；
    `POST /api/attendance/batch`、`GET /api/attendance`、`POST /api/growth/class-eval` 增加可选 `session_id`。
  version: 1.0.0
  contact:
    name: 教务管理系统
servers:
  - url: http://localhost:3000
    description: 本地 / Docker 后端
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 统一登录入口
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/info:
    get:
      tags: [Auth]
      summary: 当前用户信息（含 avatar / phone）
      description: >-
        `avatar` 为头像**相对路径**；空串表示未上传头像，前端回落内置占位图。
      responses:
        "200": { description: 成功 }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/password:
    put:
      tags: [Auth]
      summary: 自助修改密码（admin / teacher 均可改自己的）
      description: >-
        必须校验原密码；新密码至少 8 位且不能与原密码相同。成功后递增该账号的
        `users.token_version`（H2 口径），**该账号全部已签发凭证立即失效**，前端应登出并跳登录页。
        管理员重置**他人**密码仍走 `PUT /api/users/:id/password`（仅 admin）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, password]
              properties:
                old_password: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: 已修改，凭证已吊销 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/profile:
    put:
      tags: [Auth]
      summary: 自助修改个人资料（姓名 / 手机号）
      description: >-
        只能改自己的 `name` 与 `phone`。**用户名与角色不可自助修改**（用户名是登录标识、
        角色是权限边界，只能由 admin 在员工账号页改）。手机号允许传空串清空；非空时校验 11 位大陆手机号，
        与他人重复返回 400。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 50 }
                phone: { type: string }
      responses:
        "200": { description: 已更新，返回最新资料 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/avatar:
    post:
      tags: [Auth]
      summary: 自助上传头像（原始二进制直传）
      description: >-
        请求体为图片二进制（`express.raw()` 直收，不引 multer）。按**文件头魔数**判型
        （png / jpg / webp / svg / ico，忽略客户端声明的 MIME），上限 2MB（超限由 nginx 层返回 413）。
        落盘 `server/data/assets/avatars/`，文件名 `avatar-{用户id}-{时间戳}-{rand8}{ext}`；
        **DB 只存相对路径**（ADR-008），上传后自动清理该用户上一张头像。
        与站点 Logo 上传（`POST /api/site-info/upload`）共用同一份判型实现 `server/src/utils/image.js`。
      requestBody:
        required: true
        content:
          image/*:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: 已上传，返回头像相对路径
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { description: 文件过大（上限 2MB） }
  /api/auth/async-routes:
    get:
      tags: [Auth]
      summary: 动态菜单路由（按角色下发，前端自动过滤无权限菜单）
      responses:
        "200": { description: 成功 }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/refresh-token:
    post:
      tags: [Auth]
      summary: 刷新 accessToken（校验凭证未被吊销）
      description: >-
        除校验签名与过期时间外，还会比对 Token 中的 `tv` 声明与 `users.token_version`。
        该账号已登出 / 已改密 / 已改角色时返回 401。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { description: 成功 }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: 退出登录（服务端吊销该账号全部已签发凭证）
      description: >-
        递增该账号的 `users.token_version`，使其已签发的 accessToken 与 refreshToken
        立即全部失效。**需要携带有效 accessToken**（v15 起由空实现改为真实吊销）。
      responses:
        "200": { description: 已登出，该账号全部凭证已吊销 }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------------------------------------------------------- Students
...
  /api/attendance:
    get:
      tags: [Attendance]
      summary: 某课程某日全班名单（回显已登记状态）
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - { name: class_id, in: query, required: true, schema: { type: integer } }
        - { name: course_id, in: query, required: true, schema: { type: integer } }
      responses:
        "200":
          description: 名单（仅 students.status='在读'）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/AttendanceRosterItem" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/attendance/records:
    get:
      tags: [Attendance]
      summary: 考勤记录分页查询
      parameters:
        - { name: date_start, in: query, schema: { type: string, format: date } }
        - { name: date_end, in: query, schema: { type: string, format: date } }
        - { name: course_id, in: query, schema: { type: integer } }
        - { name: class_id, in: query, schema: { type: integer } }
        - { name: student_id, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [正常, 迟到, 早退, 缺勤, 请假] } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 10 } }
      responses:
        "200": { description: "分页列表（list / total）" }
  /api/attendance/statistics:
    get:
      tags: [Attendance]
      summary: 考勤统计（student / class 维度）
      parameters:
        - { name: dimension, in: query, schema: { type: string, enum: [student, class], default: student } }
        - { name: course_id, in: query, schema: { type: integer } }
        - { name: class_id, in: query, schema: { type: integer } }
        - { name: start, in: query, schema: { type: string, format: date } }
        - { name: end, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: 成功 }
...
  /api/todos:
    get:
      tags: [Todos]
      summary: 待办列表（teacher 强制仅自己；admin 可用 scope=all 看全部）
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, all], default: mine }, description: 仅 admin 的 all 生效；teacher 一律收敛为 mine }
        - { name: status, in: query, schema: { type: string, enum: [待办, 已完成] } }
        - { name: priority, in: query, schema: { type: string, enum: [普通, 重要, 紧急] } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: owner_id, in: query, schema: { type: integer }, description: 仅 admin 且 scope=all 时生效 }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200": { description: 成功（data.list / data.total / data.pendingCount） }
    post:
      tags: [Todos]
      summary: 新建待办（owner 默认自己；★ 仅 admin 可指派他人）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 100 }
                content: { type: string, maxLength: 5000 }
                priority: { type: string, enum: [普通, 重要, 紧急], default: 普通 }
                due_date: { type: string, format: date, nullable: true }
                owner_id: { type: integer, description: 仅 admin 可传；teacher 传他人 id 返回 403 }
      responses:
        "200": { description: 成功 }
        "400": { description: 标题为空或不存在的负责人 }
        "403": { description: 非 admin 试图指派给他人 }
...
  /api/todos/{id}:
    put:
      tags: [Todos]
      summary: 修改待办（含标记完成；本人或 admin）
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 100 }
                content: { type: string, maxLength: 5000 }
                status: { type: string, enum: [待办, 已完成], description: 转已完成会写 completed_at；退回时清空 }
                priority: { type: string, enum: [普通, 重要, 紧急] }
                due_date: { type: string, format: date, nullable: true }
                owner_id: { type: integer, description: 改负责人仅 admin }
      responses:
        "200": { description: 成功 }
        "400": { description: 标题为空或超长 }
        "403": { description: 既非本人也非 admin，或非 admin 试图改负责人 }
        "404": { description: 待办不存在 }
    delete:
      tags: [Todos]
      summary: 删除待办（本人或 admin）
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: 成功 }
        "403": { description: 既非本人也非 admin }
        "404": { description: 待办不存在 }

  # --------------------------------------------------------------- Analytics
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数错误或业务规则拒绝
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: 未登录 / Token 失效 / 凭证已被吊销（登出、改密、改角色后 token_version 不再匹配）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    TooManyRequests:
      description: 登录失败次数过多（默认 15 分钟内 10 次失败；`LOGIN_RATE_LIMIT_MAX` 可调）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: 无权限（教师越权 / 非 admin）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: object }
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 该学员已有报班记录，不能删除 }
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, example: admin123456 }
        type: { type: string, enum: [password, phone, wechat], default: password }
    LoginResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
            expires: { type: string, example: "2026/09/18 10:00:00" }
            avatar: { type: string }
            username: { type: string }
            nickname: { type: string }
            roles: { type: array, items: { type: string } }
            permissions: { type: array, items: { type: string } }
...
    AttendanceRosterItem:
      type: object
      properties:
        student_id: { type: integer }
        student_no: { type: string }
        name: { type: string }
        gender: { type: string }
        status:
          type: string
          nullable: true
          description: 未登记时为 null
          enum: [正常, 迟到, 早退, 缺勤, 请假, null]
        remark: { type: string, nullable: true }
        attendance_id: { type: integer, nullable: true }
...